abstraction of CGM class
