+++ py_replay_bg.egg-info/SOURCES.txt
@@ -9,7 +9,6 @@ py_replay_bg.egg-info/SOURCES.txt
 py_replay_bg.egg-info/dependency_links.txt
 py_replay_bg.egg-info/requires.txt
 py_replay_bg.egg-info/top_level.txt
-py_replay_bg/analyzer/__init__.py
 py_replay_bg/data/__init__.py
 py_replay_bg/dss/__init__.py
 py_replay_bg/dss/default_dss_handlers.py
